Moved 'FilesServiceState' to services::files::state module
