+++ src/services/files/mod.rs
@@ -1,3 +1,4 @@
 pub mod delete;
 pub mod get;
 pub mod insert;
+pub mod state;
