+++ src/state.rs
@@ -1,7 +1,7 @@
 use crate::auth::oauth::Oauth;
-use crate::endpoints::api::files::FilesServiceState;
 use crate::server::db::INTERNAL_DB;
 use crate::server_error::ServerError;
+use crate::services::files::state::FilesServiceState;
 use actix_web::web;
 use std::sync::Arc;
 use surrealdb::engine::remote::ws::{Client, Ws};
